Translate SQL array to Java array via casting
Allows users to consume SQL arrays more-or-less directly as the Scala
type they choose. Under the hood, it uses a cast to translate the
java.sql.Array object's contents to your desired type.

This is somewhat more dangerous than most of the type coercion performed
by jdub, as it's not being translated through the ResultSet object. It
may be advisable to handle possible type exceptions when using this
method.